@@ -4,8 +4,8 @@ namespace ProEventos.Persistence.Contratos
 {
     public interface IEventoPersist
     {
-        Task<Evento[]> GetAllEventosByTemaAsync(string tema, bool includePalestrantes = false);
-        Task<Evento[]> GetAllEventosAsync(bool includePalestrantes = false);
-        Task<Evento> GetEventoByIdAsync(int eventoId, bool includePalestrantes = false);
+        Task<Evento[]> GetAllEventosByTemaAsync(int userId, string tema, bool includePalestrantes = false);
+        Task<Evento[]> GetAllEventosAsync(int userId, bool includePalestrantes = false);
+        Task<Evento> GetEventoByIdAsync(int userId, int eventoId, bool includePalestrantes = false);
     }
 }
